% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{callCandidRegion}
\alias{callCandidRegion}
\title{Title}
\usage{
callCandidRegion(gr, cutoff, maxGap, minNumCR, bpWindow, verbose, parallel)
}
\arguments{
\item{parallel}{}
}
\description{
Title
}
