% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{cell_2}
\alias{cell_2}
\title{cell_2}
\format{
A data frame with 10000 rows and 5 variables (no column names):
\describe{
  \item{V1}{Chromosome}
  \item{V2}{Genomic coordinate}
  \item{V3}{Strand information}
  \item{V4}{Number of methylated reads}
  \item{V5}{Number of reads in total}
}
}
\usage{
data(cell_2)
}
\description{
This dataset is an example of a single-cell file that can be input to
\code{vmrseq::poolData} function. It contains the first 10000 CpG sites in
a random cell from mouse frontal cortext dataset published by Luo et al. (2017).
}
\examples{
data(cell_2)
head(cell_2)

}
\references{
Luo, Chongyuan et al. \emph{Single-cell methylomes identify neuronal
subtypes and regulatory elements in mammalian cortex.}. Science (New York, N.Y.)
vol. 357,6351 (2017): 600-604.
}
\keyword{datasets}
