% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{computeVar}
\alias{computeVar}
\title{Title}
\usage{
computeVar(gr, M, meanMeth, bpWindow, sparseNAdrop, parallel)
}
\arguments{
\item{M}{numeric matrix of binary single-cell methylation status. Row number
should be equal to the number of CpG sites and column number should be equal
to the number of cells.}

\item{meanMeth}{numeric vector of across-cell mean methylation. Length should
be equal to the number of CpG sites.}

\item{parallel}{}
}
\description{
Title
}
