% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{computeVarCutoff}
\alias{computeVarCutoff}
\title{Compute cutoff on smoothed variance for determining candidate regions}
\usage{
computeVarCutoff(alpha, meth, total, pars_u, pars_m, n = 1e+05)
}
\arguments{
\item{alpha}{level of significance}

\item{meth}{vector of meth read counts}

\item{total}{vector of total read counts}

\item{pars_u}{parameters used in the ZIBB distribution for unmethylated grouping}

\item{pars_m}{parameters used in the BB distribution for methylated grouping}

\item{n}{number of simulations}
}
\description{
Compute cutoff on smoothed variance for determining candidate regions
}
