% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pool.R
\name{data.pool}
\alias{data.pool}
\title{Pool single-cell file together into an HDF5-based SummarizedExperiment
object with sparse matrix representation.}
\usage{
data.pool(cellFiles, sep, writeDir, chrNames, colData, sparseNAdrop = TRUE)
}
\arguments{
\item{cellFiles}{Vector of character strings indicating single-cell file
paths you wish to pool into SE object(s). Cell files should be in BED-like
format, where the first 5 columns in each file must be: <chr>, <pos>, <strand>,
<meth_read>, <total_read>, in strict order. Each row contains info of one CpG.
Strand is assumed to be properly filpped. Cell files can be zipped ones in .gz
format.}

\item{sep}{the field separator character. Values on each line of cell file are
separated by this character.}

\item{writeDir}{A single character string indicating a folder directory where
you wish to store the processed SE object(s).}

\item{chrNames}{Single or a vector of character strings representing chromosome
names in cell files. Only chromosomes listed in \code{selectChrs} will be
processed.}

\item{colData}{A DataFrame or data.frame object containing colData for the SE
object (applied to all chromosomes).}

\item{sparseNAdrop}{Logical value indicating whether or not to use NA-dropped
sparseMatrix representation. 'NA-dropped' means replacing NAs as zeros and
then represent 0 as a very small positive value close to 0 so that the data
matrix is converted to a traditional sparseMatrix form with 0 as default
entry value. (see \code{sparseNAMatrix-class} from package \code{recommenderlab}).
Using NA-dropped sparseMatrix representation helps to save RAM during data
processing, as well as storage space saving to disk. Currently only supports
TRUE!!}
}
\description{
In each cell, sites with 0 < meth_read/total_read < 1 are removed.
}
