% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pool.R
\name{extractCoord}
\alias{extractCoord}
\title{Extract genomic coordinates of a particular chromosome in a cell file.}
\usage{
extractCoord(file, chr)
}
\arguments{
\item{chr}{}
}
\description{
Extract genomic coordinates of a particular chromosome in a cell file.
}
