% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pool.R
\name{fillNA}
\alias{fillNA}
\title{Fill NA's in missing cites per cell}
\usage{
fillNA(file, chr, pos_full, sep)
}
\arguments{
\item{file}{file path}

\item{chr}{chromosome name}

\item{pos_full}{a complete list of genomic coordinates}

\item{sep}{separator between columns}
}
\description{
Fill NA's in missing cites per cell
}
