% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDF5NAdrop2matrix.R
\name{HDF5NAdrop2matrix}
\alias{HDF5NAdrop2matrix}
\title{Coerce a NA-dropped HDF5Matrix back to matrix form}
\usage{
HDF5NAdrop2matrix(hdf5_matrix)
}
\arguments{
\item{HDF5Matrix}{an HDF5Matrix (e.g. assay output from vmrseq::data.pool)
with dropped NA values}
}
\description{
Coerce a NA-dropped HDF5Matrix back to matrix form
}
