% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTransitProbs.R
\name{plotTransitProbs}
\alias{plotTransitProbs}
\title{Plot fitted transition probabilities}
\usage{
plotTransitProbs(tp, lineSize = 0.2, plotTrain = T, pointSize = 0.2)
}
\arguments{
\item{lineSize}{size of fitted loess smooth line. Default value is 0.2.}

\item{plotTrain}{logical value indiating whether to plot training data.
Default is TRUE.}

\item{pointSize}{size of training data points. Only applicable when
`plotTrain = T`. Default value is 0.2.}
}
\description{
Plot fitted transition probabilities
}
