% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.summary.R
\name{region.summary}
\alias{region.summary}
\title{Compute regional average methylation for individual cells.}
\usage{
region.summary(SE, region_ranges, sparseNAdrop = is_sparse(assays(SE)[[1]]))
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object with one (and only one) assay that
contains *binary* methylation status of CpG sites in individual cells. In usual
analysis workflow (of vmrseq), \code{SE} should be the output of \code{vmrseq::data.pool}.}

\item{region_ranges}{\code{GRanges} object that contains genomic coordinates
of regions of interest.}

\item{sparseNAdrop}{logical value that represents whether the NA values are
droppped in the input \code{SE} object. \code{SE} objects output by
\code{vmrseq::data.pool} are NA dropped. See \code{?vmrseq::data.pool}
for details about NA-dropped representation.}
}
\value{
Returns a \code{SummarizedExperiment} object that contains the regional
average methylation per cell.
}
\description{
Compute regional average methylation for individual cells.
}
