% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{searchVMR}
\alias{searchVMR}
\title{Title}
\usage{
searchVMR(
  gr,
  CRI,
  maxGap,
  minNumVMR,
  minNumLong,
  maxNumMerge,
  tp,
  gradient,
  control,
  verbose,
  parallel
)
}
\arguments{
\item{parallel}{}
}
\description{
Title
}
