% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{smoothMF}
\alias{smoothMF}
\title{Kernel smoothing function}
\usage{
smoothMF(x, y, weights, chr, minInSpan, bpSpan, verbose, parallel)
}
\arguments{
\item{x}{vector of the x values}

\item{y}{vector of the y values}

\item{weights}{vector of the weight associated with each data point}

\item{chr}{string of the chromosome name}

\item{minInSpan}{minimum number of sites in the span}

\item{bpSpan}{base pair of the span}

\item{verbose}{logical value that indicates whether progress messages
should be printed to stdout}

\item{parallel}{logical value that indicates whether function should be
run in parallel}
}
\description{
Kernel smoothing function
}
