% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{toy.gr}
\alias{toy.gr}
\title{toy.gr}
\format{
A \code{GRanges} object.
}
\usage{
data(toy.gr)
}
\description{
This \code{GRanges} object is an example of output from \code{vmrseq::vmrseq.smooth} 
function.
}
\examples{
data(toy.gr)
toy.gr

}
\keyword{datasets}
