% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp.plot.R
\name{tp.plot}
\alias{tp.plot}
\title{Plot transition probability distribution}
\usage{
tp.plot(tp, line_size = 0.2, plot_train = TRUE, point_size = 0.2)
}
\arguments{
\item{tp}{'transitProbs' object storing information about trained transition
probabilities. Can be obtained from function `estimTransitProbs`}

\item{line_size}{size of fitted loess smooth line. Default value is 0.2.}

\item{plot_train}{logical value indiating whether to plot training data.
Default is TRUE.}

\item{point_size}{size of training data points. Only applicable when
`plot_train = T`. Default value is 0.2.}
}
\value{
A plot of the transition probability distribution
}
\description{
Plot transition probability distribution
}
\examples{
tp.plot(vmrseq:::tp0)

}
