% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pool.R
\name{extractInfo}
\alias{extractInfo}
\title{Extract and process methylation info of a particular chromosome from a cell file.}
\usage{
extractInfo(file, chr)
}
\arguments{
\item{chr}{}
}
\description{
Extract and process methylation info of a particular chromosome from a cell file.
}
