% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTransitProbs.R
\name{plotTransitProbs}
\alias{plotTransitProbs}
\title{Plot fitted transition probabilities}
\usage{
plotTransitProbs(tp, pointsize = 0.2, fitted = T, linesize = 0.2)
}
\arguments{
\item{tp}{}

\item{pointsize}{size of training data points. Default value is 0.2,}

\item{fitted}{logical value indiating whether to add fitted loess 
smooth line, default is TRUE.}

\item{linesize}{size of fitted loess smooth line, only applicable 
when `fitted = T`. Default value is 0.2.}
}
\description{
Plot fitted transition probabilities
}
