% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{toy.results}
\alias{toy.results}
\title{toy.results}
\format{
A \code{GRanges} object.
}
\usage{
data(toy.results)
}
\description{
This \code{list} object. It was generated by running the example from 
\code{vmrseq::vmrseqFit} function.
}
\examples{
data(toy.results)
toy.results

}
\keyword{datasets}
