% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmrseq.fit.R
\name{vmrseq.fit}
\alias{vmrseq.fit}
\title{Title}
\usage{
vmrseq.fit(
  gr,
  alpha = 0.05,
  maxGap = 2000,
  minNumCR = 5,
  minNumVMR = 5,
  maxNumMerge = 0,
  minNumLong = 0,
  gradient = TRUE,
  tp = NULL,
  control = vmrseq.optim.control(),
  verbose = TRUE,
  BPPARAM = bpparam()
)
}
\description{
Title
}
