% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmrseq.fit.R
\name{vmrseq.fit}
\alias{vmrseq.fit}
\title{Title}
\usage{
vmrseq.fit(
  gr,
  cutoff = 0.1,
  maxGap = 1000,
  minNumCR = 5,
  minNumVMR = 5,
  minNumLong = 0,
  gradient = TRUE,
  tp = NULL,
  control = vmrseq.optim.control(),
  verbose = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{BPPARAM}{}
}
\description{
Title
}
