% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmrseq.R
\name{vmrseq}
\alias{vmrseq}
\title{Main function for detecting and evaluating significance of VMRs in
single-cell bisulfite sequencing data.}
\usage{
vmrseq(
  SE,
  qVar = 0.1,
  penalty = 0,
  maxGap = 1000,
  minNumCR = 5,
  minNumVMR = 5,
  indvSmooth = TRUE,
  bpWindow = 2000,
  acrsSmooth = TRUE,
  bpSpan = 500,
  minInSpan = 5,
  tp = NULL,
  maxNumMerge = 0,
  minNumLong = 0,
  gradient = TRUE,
  control = vmrseq.control(),
  verbose = TRUE,
  BPPARAM = bpparam()
)
}
\arguments{
\item{SE}{SummarizedExperiment object with one and *only* one assay that
contains *binary* methylation status of CpG sites in individual cells. We
recommend using HDF5-based SummarizedExperiment object to prevent running
out of memory.}

\item{qVar}{positive scalar value between 0 and 1 that represents the
variance threshold. Default value is 0.10, which means the top 10% quantile
value of methylation variance is used as cutoff to detect candidate regions.}

\item{maxGap}{integer value representing maximum number of basepairs in
between neighboring CpGs to be included in the same VMR.}

\item{bpWindow}{}

\item{bpSpan}{a positive integer that represents the length in base pairs of
the smoothing span window if \code{smooth} is TRUE. Default value is 1000.}

\item{minInSpan}{positive integer that represents the minimum number of CpGs
in a smoothing span window if \code{smooth} is TRUE. Default value is 10.}

\item{tp}{transitProbs object that contains estimated transition
probabilities. Can be obtained by the 'estimTransitProbs' function. If
\code{tp==NULL}, internal transition probabilities in \code{vmrseq} is used.
Default is NULL.}

\item{maxNumMerge}{positive integer that represents the maximum number of
CpGs between two VMRs that can be tolerated when merging VMRs in the same
candidate region. Default is 0.}

\item{minNumLong}{positive integer that represents the minimum number of
CpGs to consider for a *long* candidate region. For fast computation use.
Default is 20. Long regions will be performed a more thorough search of
optimized prevalence value. Minimum value is the \code{minNumRegion}.}

\item{gradient}{should gradient descent be applied to optimize pi? If
set to FALSE, \code{vmrseq.control()$inits} will be used as candidate values
for pi, but no gradient descent will be performed.}

\item{control}{this sets the control parameters of the outer iterations
algorithm. The default setting is the \code{vmrseq.control} function.}

\item{verbose}{logical value that indicates whether progress messages should
be printed to stdout. Defaults value is TRUE.}

\item{BPPARAM}{a \code{BiocParallelParam} object to specify the parallel
backend. The default option is \code{BiocParallel::bpparam()} which will
automatically creates a cluster appropriate for the operating system.}
}
\value{
a list of two \code{GRanges} object that contains the results of the
inference.
}
\description{
Performs a two-step approach that (1) detects candidate regions,
and (2) detect variably methylated region(s) inside each candidate region
using a hidden Markov model.
}
