% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmrseq.smooth.R
\name{vmrseq.smooth}
\alias{vmrseq.smooth}
\title{Title}
\usage{
vmrseq.smooth(
  SE,
  residSmooth = TRUE,
  bpWindow = 2000,
  meanSmooth = FALSE,
  bpSpan = 0,
  minInSpan = 0,
  verbose = TRUE,
  BPPARAM = bpparam()
)
}
\description{
Title
}
